course update and save method ren

+++ course-up-backend/src/main/java/com/example/courseup/model/Course.java
@@ -28,7 +28,7 @@ public class Course {
     private String language;
     private String subtitles;
     private Integer totalStages;
-    private Integer totalDuration;
+    private Double totalDuration;
 
     private String imageId;
