+++ course-up-backend/src/main/java/com/example/courseup/model/DTO/UserCoursesDTO.java
@@ -13,7 +13,7 @@ public class UserCoursesDTO {
     private Long id;
     private LocalDate started_date;
     private LocalDate end_date;
-    private Integer current_duration;
+    private Double current_duration;
     private Integer current_stage;
     private Double course_point;
     private boolean is_finished;
@@ -22,7 +22,7 @@ public class UserCoursesDTO {
     private String name;
     private String instructor;
     private String description;
-    private Integer duration;
+    private Double duration;
     private Integer stage;
     private String imageId;
